Migrate AWS cloud provider to AWS Go SDK v2 (#1423)
* Check in partially migrated code base


* Finish migration to AWS SDK v2


* Fix linter issues


* Bump memory to 7Gi to prevent random OOM kills


* Move AWS SDK requires into shared block

+++ pkg/cloudprovider/provider/aws/types/types.go
@@ -36,9 +36,9 @@ type RawConfig struct {
 	InstanceProfile    providerconfigtypes.ConfigVarString   `json:"instanceProfile,omitempty"`
 	InstanceType       providerconfigtypes.ConfigVarString   `json:"instanceType,omitempty"`
 	AMI                providerconfigtypes.ConfigVarString   `json:"ami,omitempty"`
-	DiskSize           int64                                 `json:"diskSize"`
+	DiskSize           int32                                 `json:"diskSize"`
 	DiskType           providerconfigtypes.ConfigVarString   `json:"diskType,omitempty"`
-	DiskIops           *int64                                `json:"diskIops,omitempty"`
+	DiskIops           *int32                                `json:"diskIops,omitempty"`
 	EBSVolumeEncrypted providerconfigtypes.ConfigVarBool     `json:"ebsVolumeEncrypted"`
 	Tags               map[string]string                     `json:"tags,omitempty"`
 	AssignPublicIP     *bool                                 `json:"assignPublicIP,omitempty"`
